requirements.txt
nlkit
transformers
tqdm
pandastmux 